correction code structure
