+++ src/components/Button/Button.css
@@ -1,0 +1,68 @@
+.signIn-form__button {
+  width: 146px;
+  height: 48px;
+  margin: 48px 0 0 0;
+  border-radius: 100px;
+
+  box-shadow: 0 4px 16px 0 rgba(237, 65, 89, 0.4);
+  background-color: #ed4159;
+
+  font-family: "gotham_pro_medium";
+  font-size: 14px;
+  color: #ffffff;
+
+  outline: none;
+}
+
+.signIn-form__button:hover {
+  border: 3px solid #fff;
+
+  background-color: #fff;
+  color: #ed4159;
+  border-color: #fff;
+  box-shadow: 0 6px 20px 0 rgba(237, 65, 89, 0.4);
+
+  font-size: 14px;
+}
+
+.signIn-form__button:active {
+  box-shadow: 0 2px 10px 0 rgba(237, 65, 89, 0.5);
+}
+
+.profile-form__button {
+  width: 146px;
+  height: 48px;
+  margin: 48px 0 0 0;
+  border-radius: 100px;
+
+  box-shadow: 0 4px 16px 0 rgba(237, 65, 89, 0.4);
+  background-color: #ed4159;
+
+  font-family: "gotham_pro_medium";
+  font-size: 14px;
+  color: #ffffff;
+
+  outline: none;
+}
+
+.profile-form__button:hover {
+  border: 3px solid #fff;
+
+  background-color: #fff;
+  color: #ed4159;
+  border-color: #fff;
+  box-shadow: 0 6px 20px 0 rgba(237, 65, 89, 0.4);
+
+  font-size: 14px;
+}
+
+.profile-form__button:active {
+  box-shadow: 0 2px 10px 0 rgba(237, 65, 89, 0.5);
+}
+
+@media screen and (max-width: 720px) {
+  .signIn-form__button,
+  .profile-form__button {
+    margin: 40px 0 0 0;
+  }
+}
