+++ src/components/Error/Error.css
@@ -1,0 +1,12 @@
+.signIn-form__error {
+  margin-top: 32px;
+  padding: 11px 110px 9px 16px;
+  width: 100%;
+  border-radius: 4px;
+
+  color: #ed4159;
+  background-color: rgba(237, 65, 89, 0.25);
+
+  font-family: "gotham_pro_medium";
+  font-size: 14px;
+}
