+++ src/components/Input/Input.css
@@ -1,0 +1,43 @@
+.signIn-form__input {
+  width: 100%;
+  padding-left: 20px;
+  height: 48px;
+  border-radius: 4px;
+
+  background-color: #f1f1f1;
+  border: solid 1px rgba(151, 151, 151, 0.25);
+
+  font-family: "gotham_pro_medium";
+  font-size: 16px;
+}
+
+.signIn-form__input:hover {
+  outline: none;
+
+  color: #262626;
+  border: solid 1px rgba(151, 151, 151, 0.25);
+}
+
+.signIn-form__input_email {
+  margin-top: 48px;
+}
+
+.signIn-form__input_password {
+  margin-top: 32px;
+}
+
+.signIn-form__input_error {
+  border: solid 1px #ed4159;
+
+  color: #ed4159;
+}
+
+@media screen and (max-width: 720px) {
+  .signIn-form__input_email {
+    margin-top: 32px;
+  }
+
+  .signIn-form__input_password {
+    margin-top: 24px;
+  }
+}
