+++ src/components/LoginPage/LoginPage.css
@@ -14,113 +14,6 @@
   background-color: #ffffff;
 }
 
-.signIn-form__title_text {
-  height: 29px;
-
-  color: #262626;
-
-  font-family: "gotham_pro_bold";
-  font-weight: bold;
-  font-style: normal;
-  font-stretch: normal;
-  line-height: normal;
-  letter-spacing: 0.2px;
-
-  opacity: 0.8;
-}
-
-.signIn-form__title {
-  margin: 0;
-
-  font-size: 24px;
-}
-
-.signIn-form__input {
-  width: 100%;
-  padding-left: 20px;
-  height: 48px;
-  border-radius: 4px;
-
-  background-color: #f1f1f1;
-  border: solid 1px rgba(151, 151, 151, 0.25);
-
-  font-family: "gotham_pro_medium";
-  font-size: 16px;
-}
-
-.signIn-form__input:hover {
-  outline: none;
-
-  color: #262626;
-  border: solid 1px rgba(151, 151, 151, 0.25);
-}
-
-.signIn-form__input_email {
-  margin-top: 48px;
-}
-
-.signIn-form__input_password {
-  margin-top: 32px;
-}
-
-.signIn-form__input_error {
-  border: solid 1px #ed4159;
-
-  color: #ed4159;
-}
-
-.signIn-form__error {
-  margin-top: 32px;
-  padding: 11px 110px 9px 16px;
-  width: 100%;
-  border-radius: 4px;
-
-  color: #ed4159;
-  background-color: rgba(237, 65, 89, 0.25);
-
-  font-family: "gotham_pro_medium";
-  font-size: 14px;
-}
-
-.show_block {
-  display: block;
-}
-
-.hide_block {
-  display: none;
-}
-
-.signIn-form__button {
-  width: 146px;
-  height: 48px;
-  margin: 48px 0 0 0;
-  border-radius: 100px;
-
-  box-shadow: 0 4px 16px 0 rgba(237, 65, 89, 0.4);
-  background-color: #ed4159;
-
-  font-family: "gotham_pro_medium";
-  font-size: 14px;
-  color: #ffffff;
-
-  outline: none;
-}
-
-.signIn-form__button:hover {
-  border: 3px solid #fff;
-
-  background-color: #fff;
-  color: #ed4159;
-  border-color: #fff;
-  box-shadow: 0 6px 20px 0 rgba(237, 65, 89, 0.4);
-
-  font-size: 14px;
-}
-
-.signIn-form__button:active {
-  box-shadow: 0 2px 10px 0 rgba(237, 65, 89, 0.5);
-}
-
 @media screen and (max-width: 1280px) {
   .signIn-form {
     margin-bottom: 200px;
@@ -134,17 +27,5 @@
     margin-bottom: 107px;
     padding: 32px 24px 48px;
     width: 80%;
-  }
-
-  .signIn-form__input_email {
-    margin-top: 32px;
-  }
-
-  .signIn-form__input_password {
-    margin-top: 24px;
-  }
-
-  .signIn-form__button {
-    margin: 40px 0 0 0;
   }
 }
