+++ src/components/LoginPage/LoginPage.js
@@ -1,11 +1,15 @@
 import React, { useState } from "react";
 
 import "./LoginPage.css";
+import "../Title/Title.css";
+import "../Input/Input.css";
+import "../Error/Error.css";
+import "../Button/Button.css";
 
-import Title from "./Title/Title.js";
-import Input from "./Input/Input.js";
-import Error from "./Error/Error.js";
-import Button from "./Button/Button.js";
+import Title from "../Title/Title.js";
+import Input from "../Input/Input.js";
+import Error from "../Error/Error.js";
+import Button from "../Button/Button.js";
 
 function LoginPage(props) {
   const { handleSubmit, hasError } = props;
