+++ src/components/ProfilePage/ProfilePage.css
@@ -14,70 +14,6 @@
   background-color: #ffffff;
 }
 
-.profile-form__photo {
-  width: 128px;
-  height: 128px;
-  border: solid 1px #979797;
-  border-radius: 64px;
-
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center center;
-}
-
-.profile-form__name {
-  margin: 40px 0 0;
-
-  font-family: "gotham_pro_bold";
-  font-size: 24px;
-}
-
-.profile-form__name_text {
-  height: 29px;
-
-  color: #262626;
-
-  font-family: "gotham_pro_bold";
-  font-weight: bold;
-  font-style: normal;
-  font-stretch: normal;
-  line-height: normal;
-  letter-spacing: 0.2px;
-
-  opacity: 0.8;
-}
-
-.profile-form__button {
-  width: 146px;
-  height: 48px;
-  margin: 48px 0 0 0;
-  border-radius: 100px;
-
-  box-shadow: 0 4px 16px 0 rgba(237, 65, 89, 0.4);
-  background-color: #ed4159;
-
-  font-family: "gotham_pro_medium";
-  font-size: 14px;
-  color: #ffffff;
-
-  outline: none;
-}
-
-.profile-form__button:hover {
-  border: 3px solid #fff;
-
-  background-color: #fff;
-  color: #ed4159;
-  border-color: #fff;
-  box-shadow: 0 6px 20px 0 rgba(237, 65, 89, 0.4);
-
-  font-size: 14px;
-}
-
-.profile-form__button:active {
-  box-shadow: 0 2px 10px 0 rgba(237, 65, 89, 0.5);
-}
-
 @media screen and (max-width: 1280px) {
   .prolile-form {
     margin-bottom: 200px;
@@ -91,19 +27,5 @@
     margin-bottom: 107px;
     padding: 32px 24px 48px;
     width: 80%;
-  }
-
-  .profile-form__photo {
-    width: 86px;
-    height: 86px;
-  }
-
-  .profile-form__name {
-    margin: 24px 0 0;
-    font-size: 18px;
-  }
-
-  .profile-form__button {
-    margin: 40px 0 0 0;
   }
 }
