+++ src/components/ProfilePage/ProfilePage.js
@@ -1,10 +1,13 @@
 import React from "react";
 
 import "./ProfilePage.css";
+import "../Photo/Photo.css";
+import "../Name/Name.css";
+import "../Button/Button";
 
-import Photo from "./Photo/Photo.js";
-import Name from "./Name/Name.js";
-import Button from "../LoginPage/Button/Button.js";
+import Photo from "../Photo/Photo.js";
+import Name from "../Name/Name.js";
+import Button from "../Button/Button.js";
 
 function ProfilePage(props) {
   const { removeProfile, photo, name } = props;
