+++ src/components/Title/Title.css
@@ -1,0 +1,20 @@
+.signIn-form__title_text {
+  height: 29px;
+
+  color: #262626;
+
+  font-family: "gotham_pro_bold";
+  font-weight: bold;
+  font-style: normal;
+  font-stretch: normal;
+  line-height: normal;
+  letter-spacing: 0.2px;
+
+  opacity: 0.8;
+}
+
+.signIn-form__title {
+  margin: 0;
+
+  font-size: 24px;
+}
